#서영준 어드민 meeting_num Long로 변경

+++ branches/teamB_project/src/main/java/com/ync/project/admin/service/AdminMeetingService.java
@@ -13,7 +13,7 @@ import com.ync.project.domain.MeetingVO;
   */
 public interface AdminMeetingService {
 	// 글 상세보기
-	public MeetingVO read(int meeting_Num);
+	public MeetingVO read(Long meeting_Num);
 	
 	// 전체 글 목록
 	public List<MeetingVO> getList();
